clean up typography
