+++ package.json
@@ -55,6 +55,7 @@
     "@types/lodash": "^4.14.149",
     "@types/react": "^16.9.9",
     "@types/react-dom": "^16.9.2",
+    "@types/theme-ui": "^0.3.7",
     "autoprefixer": "^9.8.4",
     "css-loader": "^3.4.2",
     "docz": "^2.3.1",
