+++ packages/docz-tools/src/gatsby-theme-docz/components/Layout/index.tsx
@@ -1,19 +1,18 @@
 /** @jsx jsx */
 import { Box, jsx } from "theme-ui";
-import { Fragment } from "react";
+import { Fragment, PropsWithChildren } from "react";
 import { Global } from "@emotion/react";
-import * as styles from "./styles";
-import { global } from "../../theme/global";
+import { global } from "~theme/global";
 
 // eslint-disable-next-line import/no-relative-parent-imports
 // eslint-disable-next-line import/no-internal-modules
 import "@jobber/design/foundation.css";
 
-export function Layout({ children }) {
+export function Layout({ children }: PropsWithChildren<{}>) {
   return (
     <Fragment>
       <Global styles={global} />
-      <Box sx={styles.layout}>
+      <Box>
         <Box></Box>
         <Box>
           <Box>{children}</Box>
